+++ CHANGES.md
@@ -2,7 +2,7 @@
 
 ## Version 2.4.5 (2025/??/??)
 
-- 
+- [GH-44](https://github.com/jenkinsci/tap-plugin/pull/44): Add @Symbol("publishTap") to TapPublisher Descriptor and pipeline test (thanks @0xShubhamSolanki)
 
 ## Version 2.4.4 (2025/03/13)